+++ src/solverpy/__init__.py
@@ -1,17 +1,8 @@
 """
 
-# SolverPy Overview
-
-<div class="image-container">
-  <img src="../../diagrams/out/overview.svg" alt="Your Image" class="clickable-image">
-</div>
-"""
-
-from .setups.setup import Setup
+```dia
+overview
+```
 
 """
-This is second comment.
-and its second line..
-"""
 
-__all__ = ["Setup"]