+++ src/solverpy/task/solvertask.py
@@ -1,3 +1,7 @@
+"""
+Defines the [`SolverTask`][solverpy.task.solvertask.SolverTask] class.
+"""
+
 from typing import TYPE_CHECKING, Any
 from .task import Task
 
@@ -7,6 +11,12 @@ if TYPE_CHECKING:
 
 
 class SolverTask(Task):
+   """
+   Represents a task to be executed by a
+   [`SolverPy`][solverpy.solver.solverpy.SolverPy] solver. 
+   This is a basic interface for the database which stores results
+   of [`SolverTask`][solverpy.task.solvertask.SolverTask].
+   """
 
    def __init__(
       self,
@@ -16,6 +26,17 @@ class SolverTask(Task):
       problem: str,
       calls: list[tuple[str, str, Any, Any]] = [],
    ):
+      """
+      Create a new task to execute `solver` on problem `problem` from benchmark
+      `bid` using strategy `sid`.
+
+      Args:
+          solver: the solver based on [`SolverPy`][solverpy.solver.solverpy.SolverPy]
+          bid: the benchmark id
+          sid: the strategy id
+          problem: the problem from the benchmark
+          calls: the plugin calls to be executed before solving the task
+      """
       Task.__init__(self)
       self.solver = solver
       self.bid = bid
@@ -27,20 +48,63 @@ class SolverTask(Task):
       return f"{self.solver}:{self.sid} @ {self.bid} / {self.problem}"
 
    def run(self) -> "Result":
+      """
+      Run the task and return the result. 
+
+      First, plugin calls from `calls` given to the constructor are executed on
+      the [`solver`][solverpy.task.solvertask.SolverTask.solver]. Then,
+      [`SolverPy.solve`][solverpy.solver.solverpy.SolverPy.solve] is called on
+      [`instance`][solverpy.task.solvertask.SolverTask.instance] and
+      [`strategy`][solverpy.task.solvertask.SolverTask.strategy].
+
+      Returns: 
+         the result dictionary
+      """
       for (pid, method, args, kwargs) in self._calls:
          self.solver.call(pid, method, *args, **kwargs)
       return self.solver.solve(self.instance, self.strategy)
 
    def status(self, result: "Result") -> bool | None:
+      """
+      Translate the result to (typically smaller) status to send it over the
+      queue.
+
+      Args:
+          result: the result
+
+      Returns:
+          the status
+      """
       if not self.solver.valid(result):
          return None
       return self.solver.solved(result)
 
    @property
    def instance(self) -> tuple[str, str]:
+      """
+      Get the instance of this solver task as a benchmark-problem pair.
+
+      Returns:
+          the pair of benchmark id and problem name
+      """
       return (self.bid, self.problem)
 
    @property
    def strategy(self) -> str:
+      """
+      Get the strategy of this solver task as a strategy id.
+
+      Returns:
+          the strategy
+      """
       return self.sid
 
+   @property
+   def solver(self) -> "SolverPy":
+      """
+      Get the [`SolverPy`][solverpy.solver.solverpy.SolverPy] solver of this solver task.
+
+      Returns:
+          the solver
+      """
+      return self.solver